fix: don't wait for route broadcast message delivery
the connector was waiting for all messages to be delivered before
continuing to broadcast routes. this was discovered because
ilp-plugin-virtual's sendMessage was hanging and this was preventing the
connector from broadcasting routes. this change makes the broadcast
function synchronous. we'll fire off each message but won't wait for the
promise to finish, since all we'll do is log an error if it rejects
anyway.